Create font2mobileconfig.sh
This is a script that wraps any file in an Apple mobile provisioning profile as if that file were a font.  If you use the script on a TTF or OTF font file, you'll get a ".mobileconfig" file that can be installed on most versions of iOS or recent versions of OS X to make that font generally available in that environment.